drivers/virtio/net: Use only a single descriptor for packets
For modern virtio devices the header can be put into the same descriptor
as the data. This also now correctly removes the padded header from the
packets. The previous code added junk data at the end because it did not
consider the header part of the reported data length.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Maria Pana <maria.pana4@gmail.com>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
GitHub-Closes: #1339