lib/ukfile: Add backing buffer shared memory API
This change expands the ukfile API to provide direct shared memory
access to file backing buffers, useful for implementing memory mappings
(mmap) or single-copy file-to-file transfers (sendfile).

A boilerplate no-op implementation is provided, and existing ukfiles are
patched to use it. No actual implementations of this API are included in
this commit.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
GitHub-Closes: #1411