lib/posix-process: Fix clone() return value on thread create
Commit 08c6ac9b0ffd ("lib/posix-process: Update semantics of vfork()'s return type")
introduced a regression on the return value of clone() when new threads
are created. clone() should return the tid on the creation of a new
thread, and the pid when a new process is created. Although in the
current implementation the pid of a new process matches the underlying
tid, fix this by using the corresponding macro for each case, to keep
semantics clear.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Andrei Tatar <andrei@unikraft.io>
GitHub-Closes: #1649